Fix initial pull up controller state
This commit will add the new parameter `initialStickyPointOffset` to the method `addPullUpController` in order to provide a way to set the initial state of the pull up controller’s view.
Previusly that was not possible since the initial offset was set by `pullUpControllerPreviewOffset`.
This ticket should fix the issues #21 and #22.